Removed scripts related to not yet valid sdl sertif and HandShake expectations
