+++ test_scripts/Security/GetSystemTime/004_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_is_not_valid.lua
