+++ test_scripts/Security/GetSystemTime/006_GetSystemTime_mobile_sdl_cer_become_valid.lua
