+++ test_scripts/Security/GetSystemTime/008_GetSystemTime_mobile_cer_expired_sdl_cer_valid.lua
