+++ test_scripts/Security/GetSystemTime/012_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_becomes_not_valid.lua
