+++ test_scripts/Security/GetSystemTime/013_GetSystemTime_mobile_cer_valid_sdl_cer_not_valid.lua
@@ -7,9 +7,9 @@
 -- Description:
 -- In case:
 -- 1) Mobile app starts secure RPC service
--- 2) Mobile and sdl certificates are up to date
+-- 2) Mobile certificate is up to date and sdl certificates is not up to date
 -- 3) SDL requests GetSystemTime
--- 4) Mobile certificate and sdl certificate become not valid yet according to date/time from GetSystemTime response
+-- 4) Mobile certificate is valid and sdl certificate is not valid yet according to date/time from GetSystemTime response
 -- SDL must:
 -- 1) trigger PTU
 -- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
@@ -35,14 +35,20 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2011,
+  year = 2020,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
 local function ptUpdateWithNotActualCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0312_17.pem"
+  local crt = common.readFile(filePath)
+  pTbl.policy_table.module_config.certificate = crt
+end
+
+local function ptUpdateWithExpiredCer(pTbl)
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0312_17.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
 end
@@ -58,7 +64,7 @@ runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
 runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotActualCer })
 runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
-	{ pData, serviceId, 1, systemTime, ptUpdateWithNotActualCer })
+	{ pData, serviceId, 1, systemTime, ptUpdateWithExpiredCer })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
