+++ test_scripts/Security/GetSystemTime/018_GetSystemTime_mobile_cer_and_sdl_cer_valid.lua
