+++ test_scripts/Security/GetSystemTime/019_GetSystemTime_mobile_cer_becomes_not_valid_and_sdl_cer_valid.lua
