+++ test_scripts/Security/GetSystemTime/020_GetSystemTime_mobile_cer_becomes_yet_not_valid_and_sdl_cer_valid.lua
