+++ test_scripts/Security/GetSystemTime/022_GetSystemTime_mobile_cer_and_sdl_cer_become_expired.lua
