+++ test_scripts/Security/GetSystemTime/023_GetSystemTime_without_response_from_HMI.lua
@@ -7,11 +7,11 @@
 -- Description:
 -- In case:
 -- 1) Mobile app starts secure RPC service
--- 2) Mobile certificate is up to date and sdl certificates is not up to date
+-- 2) Mobile and sdl certificates are up to date
 -- 3) SDL requests GetSystemTime
--- 4) Mobile certificate is valid and sdl certificate is not valid yet according to date/time from GetSystemTime response
+-- 4) HMI does not respond
 -- SDL must:
--- 1) trigger PTU
+-- 1) wait default timeout
 -- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
@@ -28,27 +28,9 @@ local pData = {
   encryption = false
 }
 
-local systemTime = {
-  millisecond = 100,
-  second = 30,
-  minute = 29,
-  hour = 15,
-  day = 20,
-  month = 1,
-  year = 2020,
-  tz_hour = -3,
-  tz_minute = 10
-}
-
 --[[ Local Functions ]]
-local function ptUpdateWithNotActualCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
-  local crt = common.readFile(filePath)
-  pTbl.policy_table.module_config.certificate = crt
-end
-
-local function ptUpdateWithExpiredCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0312_17.pem"
+local function ptUpdate(pTbl)
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
 end
@@ -62,9 +44,9 @@ runner.Title("Test")
 
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
-runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotActualCer })
-runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
-	{ pData, serviceId, 1, systemTime, ptUpdateWithExpiredCer })
+runner.Step("PolicyTableUpdate with valid certificate", common.policyTableUpdate, { ptUpdate })
+runner.Step("Handshake without BC.GetSystemTime response from HMI", common.startServiceSecuredWitTimeoutWithoutGetSTResp,
+  { pData, serviceId })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
